+++ QP_solver/examples/QP_solver/solve_convex_hull_containment_lp3.h
@@ -28,7 +28,7 @@ solve_convex_hull_containment_lp (const Point_d& p,
   // construct program and solve it
   return CGAL::solve_nonnegative_linear_program
     (CGAL::make_nonnegative_linear_program_from_iterators
-     (end-begin,                                                           // n
+     (static_cast<int>(end-begin),                                         // n
       p.dimension()+1,                                                     // m
       boost::transform_iterator
       <Homogeneous_begin<Point_d>, RandomAccessIterator>(begin),           // A
